+++ CHANGES.rst
@@ -5,7 +5,10 @@ Changelog of lizard-graph
 0.14 (unreleased)
 -----------------
 
-- Upgraded views to work with lizard-fewsnorm 0.13 (still experimental).
+- Upgraded views to work with lizard-fewsnorm 0.13 (still
+  experimental).
+
+- Graphs work with timeseries and aggregated timeseries.
 
 
 0.13 (2012-03-26)